backend/requirements.txt
asgiref==3.7.2
boto3==1.34.144
botocore==1.34.144
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
contourpy==1.3.0
cryptography==42.0.2
cycler==0.12.1
defusedxml==0.7.1
dj-rest-auth==5.0.2
Django==4.0.3
django-allauth==0.60.1
django-appconf==1.0.6
django-cors-headers==4.3.1
django-cryptography==1.1
django-environ==0.11.2
django-extensions==3.2.3
django-request==1.6.3
django-storages==1.14.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
fonttools==4.53.1
gunicorn==21.2.0
idna==3.6
inflection==0.5.1
jmespath==1.0.1
kiwisolver==1.4.7
Markdown==3.5.1
matplotlib==3.9.2
numpy==2.1.1
oauthlib==3.2.2
packaging==23.2
pillow==10.4.0
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
PyMySQL==1.1.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-aws4auth==1.2.3
requests-oauthlib==1.3.1
s3transfer==0.10.2
setuptools==75.1.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
uWSGI==2.0.23
